refactor metrics repository protocol
